+++ ams/routines/uc.py
@@ -283,10 +283,10 @@ class UC(SRBase, NSRBase, MPBase, RTEDBase, DCOPF):
         # --- line ---
         self.plf.horizon = self.timeslot
         self.plf.info = '2D Line flow'
-        self.plflb.e_str = '-Bf@aBus - Pfinj - cp.multiply(rate_a, tlv) <= 0'
-        self.plfub.e_str = 'Bf@aBus + Pfinj - cp.multiply(rate_a, tlv) <= 0'
-        self.alflb.e_str = '-CftT@aBus - amax@tlv <= 0'
-        self.alfub.e_str = 'CftT@aBus - amax@tlv <= 0'
+        self.plflb.e_str = '-Bf@aBus - Pfinj - cp.multiply(ul, rate_a)@tlv <= 0'
+        self.plfub.e_str = 'Bf@aBus + Pfinj - cp.multiply(ul, rate_a)@tlv <= 0'
+        self.alflb.e_str = '-cp.multiply(ul, CftT@aBus) - amax@tlv <= 0'
+        self.alfub.e_str = 'cp.multiply(ul, CftT@aBus) - amax@tlv <= 0'
 
         # --- unserved load ---
         self.pdu = Var(info='unserved demand',